requirements.txt
python-telegram-bot==22.2
python-dotenv>=1.0.0
gigachat>=0.1.44
langchain-gigachat>=0.3.14
yookassa>=3.7.0
supabase>=2.10.0
psycopg2-binary>=2.9.9
pydantic>=2.10.0,<3.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0 
reportlab>=4.0.6
requests>=2.31.0
asyncio
flask>=3.0.0
gunicorn>=21.2.0
httpx>=0.27.0,<0.29.0 